+++ source/Energinet.DataHub.MeteringPoints.Domain/BusinessProcesses/BusinessProcessType.cs
@@ -14,7 +14,7 @@
 
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 
-namespace Energinet.DataHub.MeteringPoints.Domain
+namespace Energinet.DataHub.MeteringPoints.Domain.BusinessProcesses
 {
     public class BusinessProcessType : EnumerationType
     {
