+++ source/Energinet.DataHub.MeteringPoints.IntegrationTests/AssertBusinessProcess.cs
@@ -15,6 +15,7 @@
 using System;
 using Dapper;
 using Energinet.DataHub.MeteringPoints.Domain;
+using Energinet.DataHub.MeteringPoints.Domain.BusinessProcesses;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
 using Energinet.DataHub.MeteringPoints.Infrastructure.DataAccess;
 using Xunit;
